detector/requirements.txt
pyyaml==6.0.1
aiohttp==3.9.1
jinja2==3.1.2
psutil==5.9.6
